python_projects/ai_study/requirements.txt
sentence-transformers
faiss-cpu
hf_xet
llama-cpp-python